Add separate component for actions table
