+++ app/components/TimeNode/StatisticsComponents/index.js
@@ -1,0 +1,2 @@
+export { ExecutedGraph } from './ExecutedGraph';
+export { ActionsTable } from './ActionsTable';
