+++ app/components/TimeNode/TimeNodeStatistics.js
@@ -3,11 +3,9 @@ import { inject, observer } from 'mobx-react';
 import PropTypes from 'prop-types';
 import Alert from '../Common/Alert';
 import { TIMENODE_STATUS } from '../../stores/TimeNodeStore';
-import ExecutedGraph from './ExecutedGraph';
+import { ActionsTable, ExecutedGraph } from './StatisticsComponents';
 import { BeatLoader } from 'react-spinners';
 import moment from 'moment';
-
-const PAGE_SIZE = 100;
 
 @inject('timeNodeStore')
 @inject('keenStore')
@@ -22,14 +20,12 @@ class TimeNodeStatistics extends Component {
 
     this.state = {
       scanning: scanningStarted,
-      lastStarted: moment(),
-      currentPage: 1
+      lastStarted: moment()
     };
 
     this.startTimeNode = this.startTimeNode.bind(this);
     this.stopTimeNode = this.stopTimeNode.bind(this);
     this.refreshStats = this.refreshStats.bind(this);
-    this.changePage = this.changePage.bind(this);
     this.shouldShowClaimedWarning = this.shouldShowClaimedWarning.bind(this);
   }
 
@@ -119,12 +115,6 @@ class TimeNodeStatistics extends Component {
     );
   }
 
-  changePage(page) {
-    this.setState({
-      currentPage: page
-    });
-  }
-
   async shouldShowClaimedWarning() {
     const claimed = await this.props.timeNodeStore.getClaimedNotExecutedTransactions();
     if (claimed > 0) {
@@ -186,7 +176,7 @@ class TimeNodeStatistics extends Component {
       discovered
     } = this.props.timeNodeStore;
 
-    const { scanning, lastStarted, currentPage } = this.state;
+    const { scanning, lastStarted } = this.state;
 
     const { DISABLED, LOADING } = TIMENODE_STATUS;
     const timeNodeDisabled = nodeStatus === DISABLED || nodeStatus === LOADING;
@@ -207,28 +197,6 @@ class TimeNodeStatistics extends Component {
       ) : (
         <BeatLoader size={12} style={{ align: 'center' }} />
       );
-
-    const concatActions = actionArrays => {
-      let concatedActions = [];
-      actionArrays.forEach(actions => {
-        if (actions !== null) {
-          actions.forEach(action => concatedActions.push(action));
-        }
-      });
-      return concatedActions.sort((a, b) => b.timestamp - a.timestamp);
-    };
-
-    const allActions = concatActions([
-      successfulClaims,
-      failedClaims,
-      successfulExecutions,
-      failedExecutions
-    ]);
-
-    const pagination = { from: (currentPage - 1) * PAGE_SIZE, to: currentPage * PAGE_SIZE };
-
-    const filteredActions = allActions.slice(pagination.from, pagination.to);
-    const numPages = allActions.length > 0 ? Math.ceil(allActions.length / PAGE_SIZE) : 0;
 
     return (
       <div id="timeNodeStatistics">
@@ -452,71 +420,9 @@ class TimeNodeStatistics extends Component {
             </div>
           </div>
 
-          <div className="card-body">
-            <div className="pull-right font-montserrat">
-              Showing actions: {pagination.from}-
-              {pagination.to > allActions.length ? allActions.length : pagination.to}/
-              {allActions.length}
-            </div>
-          </div>
+          <ActionsTable />
 
-          <div className="auto-overflow" style={{ height: '250px' }}>
-            <table className="table table-condensed header-fixed">
-              <thead>
-                <tr>
-                  <th className="font-montserrat">Time</th>
-                  <th className="font-montserrat">Action</th>
-                  <th className="font-montserrat">Transaction</th>
-                  <th className="font-montserrat">Bounty</th>
-                  <th className="font-montserrat">Cost</th>
-                  <th className="font-montserrat">Outcome</th>
-                </tr>
-              </thead>
-              <tbody>
-                {filteredActions.map((action, i) => {
-                  switch (action.action) {
-                    case 1:
-                      action.action = 'Claim';
-                      break;
-                    case 2:
-                      action.action = 'Execute';
-                  }
-
-                  switch (action.result) {
-                    case 0:
-                      action.result = 'Failed';
-                      break;
-                    case 1:
-                      action.result = 'Success';
-                  }
-
-                  return (
-                    <tr key={i}>
-                      <td>{moment(action.timestamp).format('DD/MM/YYYY HH:mm:ss')}</td>
-                      <td className="font-montserrat all-caps">{action.action}</td>
-                      <td className="hint-text small">
-                        <a href="#">{action.txAddress}</a>
-                      </td>
-                      <td className="">{action.bounty} wei</td>
-                      <td className="">{action.cost} wei</td>
-                      <td className="font-montserrat all-caps b-l b-dashed b-grey">
-                        {action.result}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </div>
           <div className="padding-25">
-            {Array.from(Array(numPages).keys()).map(i => {
-              const page = i + 1;
-              return (
-                <span key={i} className="px-3" onClick={() => this.changePage(page)}>
-                  {page}
-                </span>
-              );
-            })}
             <p className="small no-margin">
               <a href="#">
                 <i className="fa fs-16 fa-arrow-circle-o-down text-success m-r-10" />
